requirements.txt
aiogram==3.15.0
anthropic
httpx
aiosqlite==0.21.0
python-dotenv==1.1.0
